2 tehtäväävaille valmis 4 osa.

+++ osa4/backend/controllers/blogs.js
@@ -22,36 +22,55 @@ blogsRouter.get('/:id', async (request, response, next) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
-  } catch (exception) {
-    next(exception)
-  }
-})
-
-/*const getTokenFrom = request => {
+const getTokenFrom = request => {
   const authorization = request.get('authorization')
   if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
     return authorization.substring(7)
   }
   return null
-}*/
+}
 
-blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body
-  console.log(request.token, 'at blogsrouter')
-
-  // const token = getTokenFrom(request)
+blogsRouter.delete('/:id', async (request, response, next) => {
+  const token = getTokenFrom(request)
 
   let decodedToken = null
 
   try {
     // eslint-disable-next-line no-undef
-    decodedToken = jwt.verify(request.token, process.env.SECRET)
-    // TÄSSÄ OLI if(!token ||...)
-    if (!request.token || !decodedToken.id) return response.status(401).json({ error: 'token missing or invalid' })
+    decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) return response.status(401).json({ error: 'token missing or invalid' })
+  } catch(exception) {
+    next(exception)
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (blogToDelete === null ) return response.status(400).json({ error: 'invalid blog ID' })
+
+  if ( blogToDelete.user.toString() === user.id.toString()) {
+    try {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } catch (exception) {
+      next(exception)
+    }
+  } else {
+    return response.status(401).json({ error: ' user can delete only it`s own blogs' })
+  }
+})
+
+blogsRouter.post('/', async (request, response, next) => {
+  const body = request.body
+
+  const token = getTokenFrom(request)
+
+  let decodedToken = null
+
+  try {
+    // eslint-disable-next-line no-undef
+    decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) return response.status(401).json({ error: 'token missing or invalid' })
   } catch(exception) {
     next(exception)
   }
